requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Django==1.11.20
idna==2.8
Naked==0.1.31
Pillow==6.1.0
pycryptodome==3.8.2
PyMySQL==0.9.3
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
shellescape==3.4.1
soupsieve==1.9.3
urllib3==1.25.3
